refactor: rename dirs for consistency
