+++ packages/components/src/ButtonGroup/ButtonGroup.tsx
@@ -5,7 +5,7 @@ import { OverrideClassName } from "~types/OverrideClassName"
 import {
   FilterButtonBase,
   FilterButtonBaseProps,
-} from "../FilterButton/_subcomponents/FilterButtonBase"
+} from "../FilterButton/subcomponents/FilterButtonBase"
 import styles from "./ButtonGroup.module.scss"
 
 const isFilterButton = (
