+++ packages/components/src/FilterMultiSelect/index.ts
@@ -1,5 +1,5 @@
 export * from "./FilterMultiSelect"
 export * from "./subcomponents"
-export * from "./provider"
+export * from "./context"
 export * from "./types"
 export * from "./utils"
