+++ packages/components/src/FilterMultiSelect/subcomponents/MenuPopup/MenuPopup.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { FocusScope } from "@react-aria/focus"
 import { useOverlay, DismissButton } from "@react-aria/overlays"
-import { useMenuTriggerContext } from "../../provider"
+import { useMenuTriggerContext } from "../../context"
 import styles from "./MenuPopup.module.scss"
 
 export interface MenuPopupProps {
